requirements.txt
typer>=0.12.0,<1.0
pyyaml>=6.0,<7.0
pytest>=7.0,<8.0
numpy>=1.24,<2.0
pillow>=10.0,<11.0
cairosvg>=2.7,<3.0
svgwrite>=1.4.3,<2.0
